+++ server/src/dao/pot/pot-delete-dao.ts
@@ -1,9 +1,15 @@
 import PotModel from '../../models/Pot';
-
+import mongoose from 'mongoose';
 
 async function deletePot(id: string) {
-    await PotModel.findByIdAndDelete(id); 
-    return true;
+    // Check if the id is a valid MongoDB ObjectId
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return null;
+    }
+    
+    const deletedPot = await PotModel.findByIdAndDelete(id);
+    
+    return deletedPot;
 }
 
 export default deletePot;
